Knockback gravity and tumble readjustments
Increased the tumble threshold from 80 units of KB to 90 and set a lower and upper cap on the hitstun gravity for all characters at 0.09 and 0.15 respectively to make comboing super floaties and super fastfallers less obnoxious and to make being comboed as them less frustrating. Additionally increased the added gravity-based vertical launch speed that applies during angles that `damage_fly_top` does not take override fall speed and gravity, which is currently angles of 45-135.