Grab during blockstun restriction removed
Currently, an attacker can't grab an opponent shielding if they grab anytime within the opponent's blockstun and shield drop frames. This effectively means if you initiate a grab with a move that is -2 or better on block, your grab will whiff. Removing this block of code will change that behavior to allow for you to grab at any point.