Charizard Dtaunt windbox size reduction
Charizard's current Dtaunt has a massive windbox that makes it an effective edge guarding tool. This coupled with the fact that it lasts for 20 frames, makes this a very non interactive edge guarding tool that is a bit ridiculous on this character. I will test in game to make sure the new windbox size and placement makes sense as I'm more or less taking an educated guess at its placement.