Man with gigantic sword arrested for swinging it too quickly
WIP, and extremely jank as I wrote this on a plane while only going off memory and a prewritten .txt file. Mostly so that hypothesis has something to chew on while I wait to get home.
Atm mostly frame data and damage adjustments, but there's a couple of angle adjustments and hitbox size/placement adjustments as well.