Add client-side latency control in arenas
This commit adds a client-side ui and control for input latency in arenas. This controls how many frames of input delay is added in arena online matches to facilitate delay-based matchmaking, rather than defaulting to the value that smash algorithmically  assigns you (which can be inconsistent and unnecessarily conservative). The ability to set it to Auto is still there, but now people with quality ethernet connections can safely lower the input latency to their liking.

Co-Authored-By: blu-dev <the.blu.dev@gmail.com>