Joker Aerial Grapple Hook Adjustments
This PR aims to rework how aerial grapple hook works, allowing for more interactive play for both the player and the opponent. Aerial Grapple Hook will no longer grab airborne opponents, but rather will spike them on the upper half of the wire OR hit them to keep them in the air on the lower half of the wire being shot out.