Backup was outdated, cant figure out how to bring back old commit. Apologies for ineptitude. This was my only backup not months outdated, might have forgotten to remove some things. Features bug fixes, logic improvements, less intrusive effects, slight tweaks, updated formatting to my limited knowledge. If something extraneous is in here please tell me to cut it out, im too tired for this.
My fault for updating it wrongly somehow. I use this on my personal setup otherwise I would not bother people by keeping it up to date. No idea if this was still in consideration but if it was please dont let this break that chance, send me a message and tell me what to do. Removed some stupid errors i noticed immediately