Reverts pellet buffs as dev told me they do not fit the balance
philosophy. I was also told not to buff his aerials but without those buffs I would have no PR. I did this as a request, feel free to tell me to bring back the pellet change or deny the pr. I wanted to start a dialogue but I do not want to offend the wrong person.