cancel window change
Now requires Greninja to be on the ground for at least 1 frame before being able to cancel the shadow sneak. This pushes the ability to cancel SS to one frame later. So, technically a miniscule nerf. This also resolves issues with the smoke's appearance, as previously the move was able to be cancelled before the smoke appeared.

The actual window of execution remains the same at 5 frames.